requirement.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
faiss-cpu==1.8.0
openai==1.40.3
sympy==1.13.2
mpmath==1.3.0
requests==2.32.3
